#199: created authentication hooks
