+++ src/webui/features/Identify.tsx
@@ -1,31 +1,9 @@
 
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-
-import johnDoe from '^/domain/authentication/johnDoe';
-import getMe from '^/domain/creator/getMe';
-import type CreatorView from '^/domain/creator/view/CreatorView';
-
-import { useAppContext } from '^/webui/contexts/module';
-import { awaitData } from '^/webui/utils/module';
+import { doIdentify } from '^/webui/hooks/module';
 
 export default function Feature()
 {
-    const navigate = useNavigate();
-    const context = useAppContext();
-
-    const getIdentity = () => getMe(johnDoe);
-
-    const setIdentity = (identity: CreatorView) =>
-    {
-        const redirectLocation = window.sessionStorage.getItem('redirect');
-
-        context.setIdentity(identity);
-
-        navigate(redirectLocation ?? '/timeline');
-    };
-
-    useEffect(() => awaitData(getIdentity, setIdentity), []);
+    doIdentify();
 
     return <>Identifying...</>;
 }
