+++ src/webui/hooks/module.ts
@@ -1,4 +1,6 @@
 
+export { default as doIdentify } from './doIdentify';
+export { default as doLogin } from './doLogin';
 export { default as useCreateComicPost } from './useCreateComicPost';
 export { default as useCreateComicReaction } from './useCreateComicReaction';
 export { default as useCreateCommentReaction } from './useCreateCommentReaction';
@@ -10,6 +12,7 @@ export { default as useEditProfile } from './useEditProfile';
 export { default as useEstablishRelation } from './useEstablishRelation';
 export { default as useExploreCreators } from './useExploreCreators';
 export { default as useExplorePosts } from './useExplorePosts';
+export { default as useLogout } from './useLogout';
 export { default as useNavigateHome } from './useNavigateHome';
 export { default as useNotifications } from './useNotifications';
 export { default as usePost } from './usePost';
@@ -20,3 +23,4 @@ export { default as useTogglePostRating } from './useTogglePostRating';
 export { default as useToggleReactionRating } from './useToggleReactionRating';
 export { default as useViewPostDetails } from './useViewPostDetails';
 export { default as useViewProfile } from './useViewProfile';
+
